Restyle homepage and make more accessible to new users
