+++ src/components/pages/HomePage.vue
@@ -11,39 +11,65 @@
             </v-row>
             <v-row v-else>
                 <v-col>
+                    <div class="mx-12">
+                        <h2>Project management</h2>
+
+                        <v-row class="mt-1 mb-6">
+                            <v-col md="12" lg="6">
+                                <v-card class="d-flex flex-column" style="height: 100%;">
+                                    <v-card-title>Add project</v-card-title>
+                                    <v-card-text class="d-flex flex-column" style="height: 100%;">
+                                        <div class="flex-grow-1">Select an empty folder and create a new project.</div>
+                                        <div class="text-center mt-2">
+                                            <v-btn color="primary" @click="createProject">
+                                                <v-icon class="mr-2">mdi-folder-plus-outline</v-icon>
+                                                Create project
+                                            </v-btn>
+                                        </div>
+                                    </v-card-text>
+                                </v-card>
+                            </v-col>
+                            <v-col md="12" lg="6">
+                                <v-card>
+                                    <v-card-title>New here? Try our demo project!</v-card-title>
+                                    <v-card-text>
+                                        <div>
+                                            If this is the first time you're using our application, we advise you to open the
+                                            demo project and discover what this application can do for you.
+                                        </div>
+                                        <div class="text-center mt-2">
+                                            <v-btn @click="openDemoProject">
+                                                Open demo project
+                                            </v-btn>
+                                        </div>
+                                    </v-card-text>
+                                </v-card>
+                            </v-col>
+                        </v-row>
+                    </div>
                     <recent-projects v-on:open-project="openProject" v-on:create-project="createProject"/>
                 </v-col>
                 <v-divider vertical></v-divider>
                 <v-col>
-                    <div style="display: flex; flex-direction: column; align-items: center;">
+                    <div class="d-flex justify-center align-center mt-12">
                         <img src="@/assets/logo.png" style="max-width: 200px;"/>
-                        <span class="logo-headline">
-                            Unipept Desktop
-                        </span>
-                        <span class="logo-subline">
-                            Version {{ version }}
-                            <a @click="releaseNotesActive = true">(What's new?)</a>
-                        </span>
-                        <span v-if="updateAvailable" class="logo-subline" @click="updateNotesActive = true">
-                            <v-icon style="position: relative; bottom: 2px;" color="success">mdi-arrow-up-bold</v-icon>
-                            <a>An update is available!</a>
-                        </span>
+                        <div class="d-flex flex-column align-center justify-center ml-4">
+                            <span class="logo-headline">
+                                Unipept Desktop
+                            </span>
+                            <span class="logo-subline">
+                                Version {{ version }}
+                            </span>
+                            <span v-if="updateAvailable" class="logo-subline" @click="updateNotesActive = true">
+                                <v-icon style="position: relative; bottom: 2px;" color="success">
+                                    mdi-arrow-up-bold
+                                </v-icon>
+                                <a>An update is available!</a>
+                            </span>
+                        </div>
                     </div>
-
-                    <div class="d-flex justify-center mt-12">
-                        <v-card>
-                            <v-card-title>
-                                New here? Try our demo project!
-                            </v-card-title>
-                            <v-card-text style="max-width: 400px;">
-                                If this is the first time you're using our application, we advise you to open the
-                                demo project and discover what this application can do for you.
-                                <div class="text-center mt-2">
-                                    <v-btn color="primary" @click="openDemoProject">Open demo project</v-btn>
-                                </div>
-                            </v-card-text>
-                        </v-card>
-                    </div>
+                    <!-- Release notes (display changes compared previous version of the application) -->
+                    <release-notes-card class="mt-12 mx-12"></release-notes-card>
                 </v-col>
             </v-row>
             <v-snackbar v-model="errorSnackbarVisible" bottom :timeout="-1" color="error">
@@ -64,7 +90,6 @@
                     </v-card-text>
                 </v-card>
             </v-dialog>
-            <release-notes-dialog v-model="releaseNotesActive"></release-notes-dialog>
             <update-notes-dialog v-model="updateNotesActive"></update-notes-dialog>
         </v-container>
     </div>
@@ -81,12 +106,10 @@ import { Tooltip } from "unipept-web-components";
 import StaticDatabaseManager from "@/logic/communication/static/StaticDatabaseManager";
 import DemoProjectManager from "@/logic/filesystem/project/DemoProjectManager";
 import ProjectVersionMismatchException from "@/logic/exception/ProjectVersionMismatchException";
-import ReleaseNotesDialog from "@/components/dialogs/ReleaseNotesDialog.vue";
+import ReleaseNotesCard from "@/components/cards/ReleaseNotesCard.vue";
 import UpdateNotesDialog from "@/components/dialogs/UpdateNotesDialog.vue";
 import GitHubCommunicator from "@/logic/communication/github/GitHubCommunicator";
 import Utils from "@/logic/Utils";
-import { OpenDialogOptions } from "electron";
-import DockerCommunicator from "@/logic/communication/docker/DockerCommunicator";
 
 const electron = require("electron");
 const { dialog } = electron.remote;
@@ -95,7 +118,7 @@ const app = electron.remote.app;
 @Component({
     components: {
         UpdateNotesDialog,
-        ReleaseNotesDialog,
+        ReleaseNotesCard,
         RecentProjects,
         Tooltip
     }
@@ -106,8 +129,6 @@ export default class HomePage extends Vue {
     private version: string = "";
     private loadingApplication: boolean = false;
     private loadingProject: boolean = false;
-    // Is the release notes dialog visible? (shows changes introduced in the current version of the app)
-    private releaseNotesActive: boolean = false;
     // Is the update notes dialog visible? (shows the changes between this version and the most recent one)
     private updateNotesActive: boolean = false;
     // Is there an update for the application available?
